validation image

+++ CHOM_BE/CHOM/wwwroot/Admin/js/Project/Create.js
@@ -15,41 +15,59 @@ $(document).ready(function () {
         }
     }
     $('.img-gt').on("change", function () {
-        var files = $('.img-gt').prop("files");
-        var url = "/Admin/Image/AddImage?handler=File";
-        formData = new FormData();
-        formData.append("File", files[0]);
-        deleteFile();
-        $.ajax({
-            url: url,
-            type: "POST",
-            dataType: "JSON",
-            data: formData,
-            contentType: false,
-            processData: false,
-            cache: false,
-            success: function (res) {
-                if (res.status == false) {
-                    console.log("error");
+        var _size = this.files[0].size;
+        var _mathSize = (_size / 1048576).toFixed(2);
+        if (_mathSize > 1) {
+            $('.txt-notification-gt').css('display', 'block');
+            $('.txt-notification-gt').text('Vui lòng upload hình dưới 1MB');
+            $('.img-gt').val('');
+        }
+        else {
+            $('.txt-notification-gt').css('display', 'none');
+            $('.txt-notification-gt').text('');
+            var files = $('.img-gt').prop("files");
+            var url = "/Admin/Image/AddImage?handler=File";
+            formData = new FormData();
+            formData.append("File", files[0]);
+            deleteFile();
+            $.ajax({
+                url: url,
+                type: "POST",
+                dataType: "JSON",
+                data: formData,
+                contentType: false,
+                processData: false,
+                cache: false,
+                success: function (res) {
+                    if (res.status == false) {
+                        console.log("error");
+                    }
+                    else {
+                        var link = "/uploadFiles/" + res.fileName;
+                        $('.uploadSingleFile').css("display", "block");
+                        $('.uploadSingleFile').attr("src", link);
+                        $('.txt-Img').val(res.fileName);
+                    }
                 }
-                else {
-                    var link = "/uploadFiles/" + res.fileName;
-                    $('.uploadSingleFile').css("display", "block");
-                    $('.uploadSingleFile').attr("src", link);
-                    $('.txt-Img').val(res.fileName);
-                    console.log("111");
-                }
-            }
-        });
+            });
+        }
     })
 
     $('.uploadMultiple').on("change", function () {
         var formData = new FormData();
-
         var listImage = $(this).get(0).files;
         for (var k = 0; k < listImage.length; k++) {
+            var _size = listImage.item(k).size;
+            var _mathSize = (_size / 1048576).toFixed(2);
+            if (_mathSize > 1) {
+                $('.txt-validation').css('display', 'block');
+                $('.txt-validation').text('Vui lòng upload hình dưới 1MB');
+                $('.uploadMultiple').val('');
+                return;
+            }
             formData.append('files', listImage.item(k));
         }
+        $('.txt-validation').css('display', 'none');
         var url = "/Admin/Image/AddMulImage?handler=files";
         $.ajax({
             url: url,
@@ -119,7 +137,4 @@ $(document).ready(function () {
             }
         });
     })
-    function testfunction() {
-        console.log("okekoekoeko")
-    }
 })
